requirements.txt
fleep
flask
rpi.GPIO
flask_socketio
flask_sqlalchemy
smbus
paho-mqtt
picamera
pillow
eventlet
gevent-websocket
